# platform/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_capability.apiKeyBody:
    properties:
      allowed_skills:
        type: string
      app_name:
        type: string
      scope:
        type: string
    required:
    - app_name
    type: object
  internal_capability.domainAgentBody:
    properties:
      code:
        type: string
      composed_skills:
        type: string
      domain:
        type: string
      name:
        type: string
      status:
        type: string
    required:
    - code
    - name
    type: object
  internal_capability.invokeBody:
    properties:
      input:
        type: string
      render_hint:
        description: card/form/chart/text
        type: string
      skill_code:
        type: string
    required:
    - input
    - skill_code
    type: object
  internal_capability.skillBody:
    properties:
      category:
        type: string
      code:
        type: string
      data_access_scope:
        type: string
      description:
        type: string
      is_public:
        type: boolean
      name:
        type: string
      prompt_template:
        type: string
      risk_level:
        type: string
      version:
        type: string
    required:
    - code
    - name
    type: object
...
info:
  contact: {}
  description: 企业 AI 原生研发平台后端（AI 驱动 需求→研发→测试→审批→发布 全流程）
  title: 智源 ANP 平台 API
  version: "1.0"
paths:
...
  /capabilities/invoke:
    post:
      consumes:
      - application/json
      parameters:
      - description: APIKey（与 Bearer 二选一）
        in: header
        name: X-Api-Key
        type: string
      - description: Bearer sk_anp_xxx
        in: header
        name: Authorization
        type: string
      - description: 调用入参(skill_code+input)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_capability.invokeBody'
      produces:
      - application/json
      responses:
        "200":
          description: result/api_key_id/caller_app
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: APIKey 无效
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权调用
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 技能不存在或不可用
          schema:
            additionalProperties: true
            type: object
        "502":
          description: 网关/上游错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 统一技能调用入口
      tags:
      - capability
  /capabilities/skills:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 已上架公共技能列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      summary: 公共技能目录
      tags:
      - capability
  /capabilities/skills/{sid}:
    get:
      parameters:
      - description: 技能ID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 技能详情
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 技能不存在
          schema:
            additionalProperties: true
            type: object
      summary: 公共技能详情
      tags:
      - capability
...
  /project-spaces/{id}/capabilities/api-keys:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: APIKey 列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出项目空间下的 APIKey
      tags:
      - capability
    post:
      consumes:
      - application/json
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: APIKey 内容(app_name)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_capability.apiKeyBody'
      produces:
      - application/json
      responses:
        "200":
          description: '{api_key,secret,note}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建 APIKey
      tags:
      - capability
  /project-spaces/{id}/capabilities/api-keys/{kid}/revoke:
    post:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: APIKey ID
        in: path
        name: kid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{id,status:revoked}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 吊销 APIKey
      tags:
      - capability
  /project-spaces/{id}/capabilities/domain-agents:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 领域 Agent 列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出领域 Agent
      tags:
      - capability
    post:
      consumes:
      - application/json
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 领域 Agent 内容(code+name)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_capability.domainAgentBody'
      produces:
      - application/json
      responses:
        "200":
          description: 创建的领域 Agent
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建领域 Agent
      tags:
      - capability
  /project-spaces/{id}/capabilities/domain-agents/{did}:
    delete:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 领域 Agent ID
        in: path
        name: did
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{id,deleted:true}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除领域 Agent
      tags:
      - capability
  /project-spaces/{id}/capabilities/skills:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 状态过滤(draft/pending_review/active/offline)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 技能列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出项目空间下的技能
      tags:
      - capability
    post:
      consumes:
      - application/json
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 技能内容(code+name)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_capability.skillBody'
      produces:
      - application/json
      responses:
        "200":
          description: 创建的技能
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建技能
      tags:
      - capability
  /project-spaces/{id}/capabilities/skills/{sid}:
    delete:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 技能ID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{id,deleted:true}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除技能
      tags:
      - capability
    put:
      consumes:
      - application/json
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 技能ID
        in: path
        name: sid
        required: true
        type: string
      - description: 技能内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_capability.skillBody'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的技能
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新技能
      tags:
      - capability
  /project-spaces/{id}/capabilities/skills/{sid}/approve:
    post:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 技能ID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{id,status:active}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 审批通过技能上架
      tags:
      - capability
  /project-spaces/{id}/capabilities/skills/{sid}/offline:
    post:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 技能ID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{id,status:offline}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 下线技能（紧急熔断）
      tags:
      - capability
  /project-spaces/{id}/capabilities/skills/{sid}/submit:
    post:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 技能ID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{id,status:pending_review}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 提交技能上架评审
      tags:
      - capability
  /project-spaces/{id}/capabilities/usage:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用量明细列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 用量明细
      tags:
      - capability
  /project-spaces/{id}/capabilities/usage/by-skill:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 按技能聚合的用量
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 按技能聚合的用量
      tags:
      - capability
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"